move to a new directory for crypto addresses validators
