+++ src/validators/__init__.py
@@ -2,10 +2,10 @@
 
 # local
 from .between import between
-from .btc_address import btc_address
 from .card import amex, card_number, diners, discover, jcb, mastercard, unionpay, visa
 from .country import calling_code, country_code, currency
 from .cron import cron
+from .crypto_addresses.btc_address import btc_address
 from .domain import domain
 from .email import email
 from .encoding import base58, base64
@@ -33,12 +33,9 @@ from .url import url
 from .utils import ValidationError, validator
 from .uuid import uuid
 
-# from .crypto_addresses import eth_address
-
 __all__ = (
     # ...
     "between",
-    # crypto_addresses
     "btc_address",
     # cards
     "amex",
