+++ src/validators/crypto_addresses/__init__.py
@@ -1,0 +1,1 @@
+"""Crypto addresses."""
