+++ tests/crypto_addresses/__init__.py
@@ -1,0 +1,1 @@
+"""Test crypto addresses."""
